Add files via upload
This is a small project that I made as a semester task and delivered to my course instructor. The name of this project is Student Help desk center. I have used visual programming , C# for mySQL database connection using Visual Studio 2022. The project performs CRUD (create, update, delete) operations on the data that is taken from user and giving user a beautiful UI(user-interface) that is made using drag and drop of elements like buttons , labels, picture box and much more!